load Janeway alterations

The data from Proceedings of the National Academy of Sciences Dec 2014, 111 (51) E5564-E5573 are available in what looks to be a relatively complete table that includes SNVs, SNPs, as well as copy number variation. This function just loads those data for convenience. It is likely that further filtering is going to be required before use.

get_janeway_alterations()

Value

a tibble

References

Data from https://www.pnas.org/highwire/filestream/618044/field_highwire_adjunct_files/7/pnas.1419260111.sd08.xlsx. See Proceedings of the National Academy of Sciences Dec 2014, 111 (51) E5564-E5573; DOI: 10.1073/pnas.1419260111.

Examples

alts = get_janeway_alterations()
head(alts)

#> # A tibble: 6 x 24
#>   Gene  Variant_Classif… Alteration Tumor_allele Reference_allele dbSNP_RS   bin
#>   <chr> <chr>            <chr>      <chr>        <chr>            <chr>    <dbl>
#> 1 PIK3… Missense_Mutati… p.E545D    T            G                rs12191…     0
#> 2 PTEN  Copy Number      Deleted    NA           NA               NA           0
#> 3 PTEN  Copy Number      Deleted    NA           NA               NA           0
#> 4 PTEN  Copy Number      Deleted    NA           NA               NA           0
#> 5 RB1   Copy Number      Deleted    NA           NA               NA           0
#> 6 RB1   Copy Number      Deleted    NA           NA               NA           0
#> # … with 17 more variables: Score_bin <chr>, Chromosome <chr>,
#> #   Start_position <dbl>, End_position <dbl>, Coverage <chr>,
#> #   Allelic_fraction <chr>, Number_of_Probes <chr>, Segment_Mean <chr>,
#> #   In_Segment <chr>, Pathways <chr>, COSMIC_overlapping_mutations <chr>,
#> #   COSMIC_total_alterations_in_gene <chr>, UniProt_Region <chr>,
#> #   Rationale <chr>, `Study ID` <chr>, Clinical_trials <chr>,
#> #   mutation.assessor <chr>

colnames(alts)

#>  [1] "Gene"                             "Variant_Classification"          
#>  [3] "Alteration"                       "Tumor_allele"                    
#>  [5] "Reference_allele"                 "dbSNP_RS"                        
#>  [7] "bin"                              "Score_bin"                       
#>  [9] "Chromosome"                       "Start_position"                  
#> [11] "End_position"                     "Coverage"                        
#> [13] "Allelic_fraction"                 "Number_of_Probes"                
#> [15] "Segment_Mean"                     "In_Segment"                      
#> [17] "Pathways"                         "COSMIC_overlapping_mutations"    
#> [19] "COSMIC_total_alterations_in_gene" "UniProt_Region"                  
#> [21] "Rationale"                        "Study ID"                        
#> [23] "Clinical_trials"                  "mutation.assessor"