load TARGET GISTIC data as summarized experiment

load TARGET GISTIC data as summarized experiment

target_gistic_se()

Examples

res = target_gistic_se()

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   `Gene Symbol` = col_character(),
#>   Cytoband = col_character()
#> )

#> See spec(...) for full column specifications.

res

#> class: SummarizedExperiment 
#> dim: 23109 88 
#> metadata(0):
#> assays(1): cn
#> rownames: NULL
#> rowData names(3): Gene Symbol Locus ID Cytoband
#> colnames(88): 0A4HLD 0A4HMC ... PAPWWC PATEEM
#> colData names(33): TARGET USI Gender ... Percent necrosis at Definitive
#>   Surgery metastatic

head(colData(res))

#> DataFrame with 6 rows and 33 columns
#>         TARGET USI      Gender        Race   Ethnicity Age at Diagnosis in Days
#>        <character> <character> <character> <character>                <numeric>
#> 0A4HLD      0A4HLD        Male          NA          NA                     7254
#> 0A4HMC      0A4HMC        Male       White          NA                    10205
#> 0A4HX8      0A4HX8        Male       White          NA                    10635
#> 0A4HXS      0A4HXS        Male       White          NA                     8470
#> 0A4HY5      0A4HY5        Male          NA          NA                     7264
#> 0A4I0Q      0A4I0Q      Female          NA          NA                     7269
#>        First Event Time to First Event in Days Vital Status
#>        <character>                   <numeric>  <character>
#> 0A4HLD    Censored                        3946        Alive
#> 0A4HMC    Censored                         377        Alive
#> 0A4HX8    Censored                        3050        Alive
#> 0A4HXS    Censored                        2948        Alive
#> 0A4HY5     Relapse                         218         Dead
#> 0A4I0Q       Death                         519         Dead
#>        Overall Survival Time in Days Year of Diagnosis Year of Last Follow Up
#>                            <numeric>         <numeric>              <numeric>
#> 0A4HLD                          3946              2000                   2011
#> 0A4HMC                           377                NA                     NA
#> 0A4HX8                          1750              2003                   2011
#> 0A4HXS                          2948              2004                   2014
#> 0A4HY5                           290              2004                   2005
#> 0A4I0Q                           519              2005                   2006
#>           Protocol       Disease at diagnosis Metastasis site
#>        <character>                <character>     <character>
#> 0A4HLD        IHRT Non-metastatic (confirmed)              NA
#> 0A4HMC        IHRT Non-metastatic (confirmed)              NA
#> 0A4HX8        IHRT Non-metastatic (confirmed)              NA
#> 0A4HXS        IHRT Non-metastatic (confirmed)              NA
#> 0A4HY5        IHRT                 Metastatic       Lung only
#> 0A4I0Q        IHRT                 Metastatic       Lung only
#>        Primary tumor site Specific tumor site Specific tumor side
#>               <character>         <character>         <character>
#> 0A4HLD           Leg/Foot               Tibia                  NA
#> 0A4HMC           Leg/Foot              Fibula                  NA
#> 0A4HX8           Leg/Foot               Femur                  NA
#> 0A4HXS           Leg/Foot               Femur                  NA
#> 0A4HY5           Leg/Foot              Pelvis                  NA
#> 0A4I0Q           Leg/Foot               Femur                  NA
#>        Specific tumor region Definitive Surgery Primary site progression
#>                  <character>        <character>              <character>
#> 0A4HLD              Proximal       Limb sparing                       No
#> 0A4HMC  Other, Not specified         Amputation                       No
#> 0A4HX8                Distal       Limb sparing                       No
#> 0A4HXS                Distal       Limb sparing                       No
#> 0A4HY5      Other, Posterior       Limb sparing                       No
#> 0A4I0Q                Distal       Limb sparing                       No
#>        Site of initial relapse Time to first relapse in days
#>                    <character>                     <numeric>
#> 0A4HLD                      NA                            NA
#> 0A4HMC                      NA                            NA
#> 0A4HX8                      NA                            NA
#> 0A4HXS                      NA                            NA
#> 0A4HY5    Regional Lymph nodes                           218
#> 0A4I0Q                      NA                            NA
#>        Time to first enrollment on relapse protocol in days
#>                                                   <numeric>
#> 0A4HLD                                                   NA
#> 0A4HMC                                                   NA
#> 0A4HX8                                                   NA
#> 0A4HXS                                                   NA
#> 0A4HY5                                                   NA
#> 0A4I0Q                                                   NA
#>        Time to first SMN in days Time to death in days Histologic response
#>                        <numeric>             <numeric>         <character>
#> 0A4HLD                        NA                    NA           Stage 3/4
#> 0A4HMC                        NA                    NA           Stage 1/2
#> 0A4HX8                        NA                    NA           Stage 3/4
#> 0A4HXS                        NA                    NA           Stage 1/2
#> 0A4HY5                        NA                   290           Stage 1/2
#> 0A4I0Q                        NA                   519           Stage 1/2
#>        Percent necrosis Relapse Type
#>             <character>  <character>
#> 0A4HLD               99           NA
#> 0A4HMC               60           NA
#> 0A4HX8              100           NA
#> 0A4HXS               40           NA
#> 0A4HY5               50           NA
#> 0A4I0Q             < 50           NA
#>                                                                    Therapy
#>                                                                <character>
#> 0A4HLD                                                MTX/Adria/Ifos/±Plat
#> 0A4HMC                                                          Adria/Plat
#> 0A4HX8 Pre-op Adria/Ifos/MTX + Adria/Plat; Post-op Adria/Plat + Adria/Ifos
#> 0A4HXS                                                      Adria/Plat/MTX
#> 0A4HY5                                                      Adria/Plat/MTX
#> 0A4I0Q                                                MTX/Adria/Ifos/±Plat
#>            Comment      cohort Percent necrosis at Definitive Surgery
#>        <character> <character>                            <character>
#> 0A4HLD          NA   Discovery                                     NA
#> 0A4HMC          NA   Discovery                                     NA
#> 0A4HX8          NA   Discovery                                     NA
#> 0A4HXS          NA   Discovery                                     NA
#> 0A4HY5          NA   Discovery                                     NA
#> 0A4I0Q          NA   Discovery                                     NA
#>        metastatic
#>         <logical>
#> 0A4HLD      FALSE
#> 0A4HMC      FALSE
#> 0A4HX8      FALSE
#> 0A4HXS      FALSE
#> 0A4HY5       TRUE
#> 0A4I0Q       TRUE

assay(res,'cn')[1:5,1:5]

#>      0A4HLD 0A4HMC 0A4HX8 0A4HXS 0A4HY5
#> [1,] -0.498   0.27 -0.465  0.203   0.11
#> [2,] -0.498   0.27 -0.465  0.203   0.11
#> [3,] -0.498   0.27 -0.465  0.203   0.11
#> [4,] -0.498   0.27 -0.465  0.203   0.11
#> [5,] -0.498   0.27 -0.465  0.203   0.11