load TARGET osteosarcoma data

This loads the data from the local system.file file. This file may or may not be the latest version. The current verion is from 20180107.

target_load_clinical()

Details

The clinical columns include:

TARGET USI, Gender, Race, Ethnicity, Age at Diagnosis in Days, First Event, Time to First Event in Days, Vital Status, Overall Survival Time in Days, Year of Diagnosis, Year of Last Follow Up, Protocol, Disease at diagnosis, Metastasis site, Primary tumor site, Specific tumor site, Specific tumor side, Specific tumor region, Definitive Surgery, Primary site progression, Site of initial relapse, Time to first relapse in days, Time to first enrollment on relapse protocol in days, Time to first SMN in days, Time to death in days, Histologic response, Percent necrosis at Definitive Surgery, Relapse Type, Therapy, Comment, cohort

References

ftp://caftpd.nci.nih.gov/pub/OCG-DCC/TARGET/OS/clinical/harmonized/

Examples

target_load_clinical()

#> # A tibble: 308 x 33
#>    `TARGET USI` Gender Race  Ethnicity `Age at Diagnos… `First Event`
#>    <chr>        <chr>  <chr> <chr>                <dbl> <chr>        
#>  1 TARGET-40-P… Male   Asian Not Hisp…            7270  No event     
#>  2 TARGET-40-P… Female Unkn… Not Hisp…            3942  Relapse      
#>  3 TARGET-40-P… Male   White Not Hisp…            4249  Relapse      
#>  4 TARGET-40-P… Male   White Not Hisp…            5996. Relapse      
#>  5 TARGET-40-P… Female Asian Not Hisp…            4987  Relapse      
#>  6 TARGET-40-P… Male   White Not Hisp…            6085. Relapse      
#>  7 TARGET-40-P… Male   White Not Hisp…            4889  Relapse      
#>  8 TARGET-40-P… Female Unkn… Hispanic…            7631. Relapse      
#>  9 TARGET-40-P… Male   Blac… Not Hisp…            2803  Relapse      
#> 10 TARGET-40-P… Female White Not Hisp…            4509  Relapse      
#> # ... with 298 more rows, and 27 more variables: `Time to First Event in
#> #   Days` <dbl>, `Vital Status` <chr>, `Overall Survival Time in Days` <dbl>,
#> #   `Year of Diagnosis` <dbl>, `Year of Last Follow Up` <dbl>, Protocol <chr>,
#> #   `Disease at diagnosis` <chr>, `Metastasis site` <chr>, `Primary tumor
#> #   site` <chr>, `Specific tumor site` <chr>, `Specific tumor side` <chr>,
#> #   `Specific tumor region` <chr>, `Definitive Surgery` <chr>, `Primary site
#> #   progression` <chr>, `Site of initial relapse` <chr>, `Time to first relapse
#> #   in days` <dbl>, `Time to first enrollment on relapse protocol in
#> #   days` <dbl>, `Time to first SMN in days` <dbl>, `Time to death in
#> #   days` <dbl>, `Histologic response` <chr>, `Percent necrosis` <chr>,
#> #   `Relapse Type` <chr>, Therapy <chr>, Comment <chr>, cohort <chr>, `Percent
#> #   necrosis at Definitive Surgery` <chr>, metastatic <lgl>