load COSMIC gene census

The COSMIC gene census list is just the genes that the COSMIC project has curated. These are all bona fide cancer genes. The csv file was downloaded offline, since signin is required.

cosmic_census()

References

See https://cancer.sanger.ac.uk/census

Examples

cosmic_census()

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   `Entrez GeneId` = col_double(),
#>   Tier = col_double()
#> )

#> See spec(...) for full column specifications.

#> # A tibble: 719 x 20
#>    `Gene Symbol` Name  `Entrez GeneId` `Genome Locatio…  Tier Hallmark
#>    <chr>         <chr>           <dbl> <chr>            <dbl> <chr>   
#>  1 A1CF          APOB…           29974 10:50806729-508…     2 NA      
#>  2 ABI1          abl-…           10006 10:26748570-268…     1 Yes     
#>  3 ABL1          v-ab…              25 9:130835447-130…     1 Yes     
#>  4 ABL2          c-ab…              27 1:179107718-179…     1 NA      
#>  5 ACKR3         atyp…           57007 2:236580466-236…     1 Yes     
#>  6 ACSL3         acyl…            2181 2:222908773-222…     1 Yes     
#>  7 ACSL6         acyl…           23305 5:131954234-132…     2 NA      
#>  8 ACVR1         acti…              90 2:157737531-157…     1 Yes     
#>  9 ACVR2A        acti…              92 2:147845153-147…     1 NA      
#> 10 AFF1          AF4/…            4299 4:87007409-8713…     1 Yes     
#> # ... with 709 more rows, and 14 more variables: `Chr Band` <chr>,
#> #   Somatic <chr>, Germline <chr>, `Tumour Types(Somatic)` <chr>, `Tumour
#> #   Types(Germline)` <chr>, `Cancer Syndrome` <chr>, `Tissue Type` <chr>,
#> #   `Molecular Genetics` <chr>, `Role in Cancer` <chr>, `Mutation Types` <chr>,
#> #   `Translocation Partner` <chr>, `Other Germline Mut` <chr>, `Other
#> #   Syndrome` <chr>, Synonyms <chr>