Load TARGET huex data as SummarizedExperiment

Load TARGET huex data as SummarizedExperiment

target_huex_se()

Examples

res = target_huex_se()

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   gene_assignment_final = col_character()
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   Date = col_date(format = ""),
#>   `Comment[Array Lot No]` = col_logical(),
#>   `Comment[Scanning Station No]` = col_double(),
#>   `Comment[OCG Data Level]` = col_double(),
#>   `Comment[OCG Data Level]_1` = col_double(),
#>   `Comment[OCG Data Level]_2` = col_double()
#> )

#> See spec(...) for full column specifications.

res

#> class: SummarizedExperiment 
#> dim: 22011 89 
#> metadata(0):
#> assays(1): exprs
#> rownames(22011): 3948543 3063807 ... 2968418 3275568
#> rowData names(3): symbol tx_list symbol_list
#> colnames(89): PATKSS PAUTWB ... 0A4HMC PASUUH
#> colData names(33): TARGET USI Gender ... Percent necrosis at Definitive
#>   Surgery metastatic

head(colData(res))

#> DataFrame with 6 rows and 33 columns
#>         TARGET USI      Gender                      Race              Ethnicity
#>        <character> <character>               <character>            <character>
#> PATKSS      PATKSS        Male                     White Not Hispanic or Latino
#> PAUTWB      PAUTWB        Male                     White Not Hispanic or Latino
#> PAUTYB      PAUTYB      Female                     White     Hispanic or Latino
#> PAUUML      PAUUML        Male                   Unknown     Hispanic or Latino
#> PAUVUL      PAUVUL      Female Black or African American Not Hispanic or Latino
#> PAUXPZ      PAUXPZ      Female                   Unknown                Unknown
#>        Age at Diagnosis in Days First Event Time to First Event in Days
#>                       <numeric> <character>                   <numeric>
#> PATKSS                     4889     Relapse                          63
#> PAUTWB                     5064     Relapse                         598
#> PAUTYB                     4418     Relapse                         527
#> PAUUML                     2177    No event                         672
#> PAUVUL                     2027     Relapse                         237
#> PAUXPZ                     4507     Relapse                         419
#>        Vital Status Overall Survival Time in Days Year of Diagnosis
#>         <character>                     <numeric>         <numeric>
#> PATKSS         Dead                           758              2010
#> PAUTWB        Alive                           687              2012
#> PAUTYB         Dead                           537              2012
#> PAUUML        Alive                           672              2012
#> PAUVUL         Dead                           347              2012
#> PAUXPZ         Dead                           606              2012
#>        Year of Last Follow Up    Protocol       Disease at diagnosis
#>                     <numeric> <character>                <character>
#> PATKSS                   2012    AOST0331 Non-metastatic (confirmed)
#> PAUTWB                   2014    AOST06B1 Non-metastatic (confirmed)
#> PAUTYB                   2013    AOST06B1 Non-metastatic (confirmed)
#> PAUUML                   2014    AOST06B1 Non-metastatic (confirmed)
#> PAUVUL                   2013    AOST06B1     Metastatic (confirmed)
#> PAUXPZ                   2014    AOST06B1     Metastatic (confirmed)
#>        Metastasis site Primary tumor site Specific tumor site
#>            <character>        <character>         <character>
#> PATKSS              NA             Pelvis     Pelvis / Sacrum
#> PAUTWB              NA           Leg/foot               Tibia
#> PAUTYB              NA           Leg/foot               Tibia
#> PAUUML              NA           Leg/foot               Femur
#> PAUVUL       Lung only           Leg/foot               Femur
#> PAUXPZ       Lung only           Leg/foot               Tibia
#>        Specific tumor side Specific tumor region Definitive Surgery
#>                <character>           <character>        <character>
#> PATKSS                  NA                    NA       Limb sparing
#> PAUTWB                  NA                    NA                 NA
#> PAUTYB                  NA                    NA                 NA
#> PAUUML                  NA                    NA                 NA
#> PAUVUL                  NA                    NA                 NA
#> PAUXPZ                  NA                    NA                 NA
#>        Primary site progression Site of initial relapse
#>                     <character>             <character>
#> PATKSS                      Yes                      NA
#> PAUTWB                       NA                      NA
#> PAUTYB                       NA                      NA
#> PAUUML                       NA                      NA
#> PAUVUL                       NA                      NA
#> PAUXPZ                       NA                      NA
#>        Time to first relapse in days
#>                            <numeric>
#> PATKSS                            63
#> PAUTWB                           598
#> PAUTYB                           527
#> PAUUML                            NA
#> PAUVUL                           237
#> PAUXPZ                           419
#>        Time to first enrollment on relapse protocol in days
#>                                                   <numeric>
#> PATKSS                                                   NA
#> PAUTWB                                                   NA
#> PAUTYB                                                   NA
#> PAUUML                                                   NA
#> PAUVUL                                                   NA
#> PAUXPZ                                                   NA
#>        Time to first SMN in days Time to death in days Histologic response
#>                        <numeric>             <numeric>         <character>
#> PATKSS                        NA                   758                0-90
#> PAUTWB                        NA                    NA                  NA
#> PAUTYB                        NA                   537                  NA
#> PAUUML                        NA                    NA                  NA
#> PAUVUL                        NA                   347                  NA
#> PAUXPZ                        NA                   724                  NA
#>        Percent necrosis Relapse Type     Therapy     Comment      cohort
#>             <character>  <character> <character> <character> <character>
#> PATKSS               NA           NA          NA          NA   Discovery
#> PAUTWB               NA           NA          NA          NA   Discovery
#> PAUTYB               NA           NA          NA          NA   Discovery
#> PAUUML               NA           NA          NA          NA   Discovery
#> PAUVUL               NA           NA          NA          NA   Discovery
#> PAUXPZ               NA           NA          NA          NA   Discovery
#>        Percent necrosis at Definitive Surgery metastatic
#>                                   <character>  <logical>
#> PATKSS                                     NA      FALSE
#> PAUTWB                                     NA      FALSE
#> PAUTYB                                     NA      FALSE
#> PAUUML                                     NA      FALSE
#> PAUVUL                                     NA       TRUE
#> PAUXPZ                                     NA       TRUE

assay(res,'exprs')[1:5,1:5]

#>           PATKSS  PAUTWB  PAUTYB   PAUUML  PAUVUL
#> 3948543  8.28067 8.39749 8.23240  9.34217 7.71883
#> 3063807  8.24085 8.42118 8.06178  8.04157 8.01377
#> 3096575 10.53751 9.82508 9.80597  9.32693 9.23512
#> 2949118  7.39632 7.84181 7.37891  7.28287 7.70401
#> 3227645 10.20303 9.83912 9.80234 10.56681 9.08011

head(rowData(res))

#> DataFrame with 6 rows and 3 columns
#>              symbol                                             tx_list
#>         <character>                                              <List>
#> 3948543     FAM118A          ENST00000216214,NM_001104595,NM_017911,...
#> 3063807        GPC2 ENST00000471717,ENST00000480087,ENST00000490629,...
#> 3096575      HGSNAT       NM_152419,ENST00000458501,ENST00000379644,...
#> 2949118         LTB ENST00000457552,ENST00000456113,ENST00000446745,...
#> 3227645        UCK1          NM_001261451,NM_031432,ENST00000372215,...
#> 4014076         HDX             NM_001177479,NM_144657,NM_001177478,...
#>                              symbol_list
#>                                   <List>
#> 3948543                          FAM118A
#> 3063807                             GPC2
#> 3096575 HGSNAT,METTL21A,LOC100653153,...
#> 2949118                              LTB
#> 3227645                             UCK1
#> 4014076                              HDX